api para obetner modulos por menu, validaciones y correciones en los recursos v0.1.1

+++ src/views/Client/ClientScreen.tsx
@@ -1,7 +1,6 @@
 import { Button, Card, Table } from "react-bootstrap";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useContext, useEffect, useState } from "react";
 import { Client } from "../../interface/Client";
-import { Resources } from "../../interface/Resources";
 import {
   deleteClient,
   getClients,
@@ -12,12 +11,34 @@ import styles from "./Client.module.scss";
 import ClientForm from "../../components/ClientComponent/Form/ClientForm";
 import ClientListActives from "../../components/ClientComponent/List/Actives/ClientListActives";
 import ClientListRemoves from "../../components/ClientComponent/List/Removes/ClientListRemoves";
+import { AuthContext } from "../../context/auth";
+import { useLocation } from "react-router-dom";
+import { getModuleByMenu } from "../../api/module/module";
+import { IAlert } from "../../interface/IAlert";
 
-const ClientScreen = ({ myResource }: { myResource: Resources }) => {
+const initialState: IAlert = {
+  type: "",
+  message: "",
+};
+
+const ClientScreen = () => {
   const [show, setShow] = useState(false);
   const [clients, setClients] = useState<Client[]>([]);
   const [removes, setRemoves] = useState<Client[]>([]);
   const [state, setState] = useState<any>();
+  const { resources } = useContext(AuthContext);
+  const [resource, setResource] = useState<any>(null);
+  const location = useLocation();
+  const getNameLocation = location.pathname.slice(1);
+  const [message, setMessage] = useState<IAlert>(initialState);
+
+  const getMyModule = useCallback(async () => {
+    const mymodule = await getModuleByMenu(getNameLocation);
+    const findResource = resources.find(
+      (res: any) => res.module.name === mymodule.data.name
+    );
+    setResource(findResource);
+  }, [resources, getNameLocation]);
 
   const listClients = useCallback(async () => {
     const res = await getClients();
@@ -45,6 +66,13 @@ const ClientScreen = ({ myResource }: { myResource: Resources }) => {
 
   const deleteCli = useCallback(
     async (id: string) => {
+      if (resource && resource.canDelete === false) {
+        setMessage({
+          type: "danger",
+          message: "No tienes acceso a este recurso.",
+        });
+        return;
+      }
       const __deletedClient = await deleteClient(id);
       const { data } = __deletedClient;
       const { clientDeleted } = data;
@@ -53,11 +81,18 @@ const ClientScreen = ({ myResource }: { myResource: Resources }) => {
         listClientsDeleted();
       }
     },
-    [listClients, listClientsDeleted]
+    [listClients, listClientsDeleted, resource]
   );
 
   const restoreCli = useCallback(
     async (id: string) => {
+      if (resource && resource.canRestore === false) {
+        setMessage({
+          type: "danger",
+          message: "No tienes acceso a este recurso.",
+        });
+        return;
+      }
       const __restoreCli = await restoreClient(id);
       const { data } = __restoreCli;
       const { clientRestored } = data;
@@ -66,74 +101,111 @@ const ClientScreen = ({ myResource }: { myResource: Resources }) => {
         listClientsDeleted();
       }
     },
-    [listClients, listClientsDeleted]
+    [listClients, listClientsDeleted, resource]
   );
 
   useEffect(() => {
-    listClients();
+    if (resource && resource.canRead) {
+      listClients();
+    }
+
     listClientsDeleted();
-  }, [listClients, listClientsDeleted]);
+    getMyModule();
+  }, [listClients, listClientsDeleted, getMyModule, resource]);
 
   return (
     <>
-      <ClientForm
-        show={show}
-        closeModal={closeModal}
-        listClients={listClients}
-        client={state}
-      />
-
       <Card>
         <Card.Header as="h5">Lista de Clientes</Card.Header>
         <Card.Body>
-          <Button
-            type="button"
-            variant="primary"
-            onClick={() => openModalRE(false)}
-          >
-            Agregar nuevo cliente
-          </Button>
-          <Table
-            striped
-            bordered
-            hover
-            responsive="sm"
-            className={styles.table}
-          >
-            <thead>
-              <tr>
-                <th>#</th>
-                <th>Nombres</th>
-                <th>Apellidos</th>
-                <th>Documento</th>
-                <th>Nro. de documento</th>
-                <th>Correo</th>
-                <th>Celular</th>
-                <th>Dirección</th>
-                <th className={`${styles["table--center"]}`}>Estado</th>
-                <th className={`${styles["table--center"]}`}>Eliminar</th>
-              </tr>
-            </thead>
-            <tbody>
-              {clients.map((client) => (
-                <ClientListActives
-                  key={client._id}
-                  client={client}
-                  deleteCli={deleteCli}
-                  openModalRE={openModalRE}
-                />
-              ))}
-            </tbody>
-            <tfoot>
-              {removes.map((remove) => (
-                <ClientListRemoves
-                  key={remove._id}
-                  remove={remove}
-                  restoreCli={restoreCli}
-                />
-              ))}
-            </tfoot>
-          </Table>
+          {resource && resource.canCreate && resource.canUpdate ? (
+            <>
+              <Button
+                type="button"
+                variant="primary"
+                onClick={() => openModalRE(false)}
+              >
+                Agregar nuevo cliente
+              </Button>
+              <ClientForm
+                show={show}
+                closeModal={closeModal}
+                listClients={listClients}
+                client={state}
+              />
+            </>
+          ) : resource && resource.canCreate ? (
+            <>
+              <Button
+                type="button"
+                variant="primary"
+                onClick={() => openModalRE(false)}
+              >
+                Agregar nuevo cliente
+              </Button>
+              <ClientForm
+                show={show}
+                closeModal={closeModal}
+                listClients={listClients}
+                client={state}
+              />
+            </>
+          ) : (
+            resource &&
+            resource.canUpdate && (
+              <ClientForm
+                show={show}
+                closeModal={closeModal}
+                listClients={listClients}
+                client={state}
+              />
+            )
+          )}
+          {resource && resource.canRead && (
+            <Table
+              striped
+              bordered
+              hover
+              responsive="sm"
+              className={styles.table}
+            >
+              <thead>
+                <tr>
+                  <th>#</th>
+                  <th>Nombres</th>
+                  <th>Apellidos</th>
+                  <th>Documento</th>
+                  <th>Nro. de documento</th>
+                  <th>Correo</th>
+                  <th>Celular</th>
+                  <th>Dirección</th>
+                  <th className={`${styles["table--center"]}`}>Estado</th>
+                  {resource && resource.canDelete && (
+                    <th className={`${styles["table--center"]}`}>Eliminar</th>
+                  )}
+                </tr>
+              </thead>
+              <tbody>
+                {clients.map((client) => (
+                  <ClientListActives
+                    key={client._id}
+                    client={client}
+                    deleteCli={deleteCli}
+                    openModalRE={openModalRE}
+                  />
+                ))}
+              </tbody>
+              <tfoot>
+                {removes.map((remove) => (
+                  <ClientListRemoves
+                    key={remove._id}
+                    remove={remove}
+                    restoreCli={restoreCli}
+                  />
+                ))}
+              </tfoot>
+            </Table>
+          )}
         </Card.Body>
       </Card>
     </>
